requirements.txt
Django==3.2.7
social-auth-app-django==5.0.0
django-jazzmin==2.4.8
Pillow==8.3.2
django-tinymce==3.3.0